commit_extraction/pyproject.toml
[project]
name = "commit-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daft>=0.4.7",
    "gitpython>=3.1.44",
]
